+++ src/inspect_ai/_view/www/src/samples/SamplesTools.mjs
@@ -3,6 +3,7 @@ import { html } from "htm/preact";
 import { EpochFilter } from "./tools/EpochFilter.mjs";
 import { SortFilter } from "./tools/SortFilter.mjs";
 import { SampleFilter } from "./tools/SampleFilter.mjs";
+import { SelectScorer } from "./tools/SelectScorer.mjs";
 
 export const SampleTools = (props) => {
   const {
@@ -14,10 +15,24 @@ export const SampleTools = (props) => {
     setSort,
     epochs,
     sampleDescriptor,
+    score,
+    setScore,
+    scores,
   } = props;
 
   const hasEpochs = epochs > 1;
   const tools = [];
+
+  if (scores.length > 1) {
+    tools.push(
+      html`<${SelectScorer}
+        scores=${scores}
+        score=${score}
+        setScore=${setScore}
+      />`,
+    );
+  }
+
   if (hasEpochs) {
     tools.push(
       html`<${EpochFilter}
@@ -37,7 +52,12 @@ export const SampleTools = (props) => {
   );
 
   tools.push(
-    html`<${SortFilter} sort=${sort} setSort=${setSort} epochs=${hasEpochs} />`,
+    html`<${SortFilter}
+      sampleDescriptor=${sampleDescriptor}
+      sort=${sort}
+      setSort=${setSort}
+      epochs=${hasEpochs}
+    />`,
   );
 
   return tools;