+++ tests/test_log_dir/example_task/example_task.py
@@ -1,0 +1,12 @@
+from inspect_ai import Task, task
+from inspect_ai.dataset import Sample
+from inspect_ai.scorer import match
+
+
+@task
+def example_task() -> Task:
+    task = Task(
+        dataset=[Sample(input="Say Hello", target="Hello")],
+        scorer=match(),
+    )
+    return task