+++ src/inspect_ai/model/_providers/providers.py
@@ -46,7 +46,7 @@ def anthropic() -> type[ModelAPI]:
 def google() -> type[ModelAPI]:
     FEATURE = "Google API"
     PACKAGE = "google-generativeai"
-    MIN_VERSION = "0.4.0"
+    MIN_VERSION = "0.5.3"
 
     # verify we have the package
     try: